requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.1
sqlparse==0.5.0
typing-extensions==4.11.0
urllib3==2.2.1